# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: API для Emplacc.
  title: Emplacc API
  version: "1.0"
paths:
  /admin/llm-settings:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Получить настройки LLM
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Обновить настройки LLM
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"